feat(work): 新增上下游规则相关接口

+++ src/SKIT.FlurlHttpClient.Wechat.Work/Models/CgibinKf/Servicer/CgibinKfServicerDeleteRequest.cs
@@ -26,6 +26,6 @@ namespace SKIT.FlurlHttpClient.Wechat.Work.Models
         /// </summary>
         [Newtonsoft.Json.JsonProperty("department_id_list")]
         [System.Text.Json.Serialization.JsonPropertyName("department_id_list")]
-        public IList<int>? DepartmentIdList { get; set; }
+        public IList<long>? DepartmentIdList { get; set; }
     }
 }
